requirements.txt
streamlit
torch
openai-whisper
gTTS
langid
ffmpeg-python
